pyproject.toml
[tool.poetry]
name = "calendar-audit"
version = "0.1.0"
description = ""
authors = ["Brandon <BrandonWL@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1"
djangorestframework = "^3.13.1"
django-environ = "^0.9.0"
psycopg2 = "^2.9.3"
google-api-python-client = "^2.57.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^0.5.2"
requests = "^2.28.1"
django-cors-headers = "^3.13.0"
PyJWT = "^2.4.0"
channels = "^3.0.5"

[tool.poetry.dev-dependencies]
black = "^22.6.0"
types-requests = "^2.28.9"
pre-commit = "^2.20.0"
mypy = "^0.971"
django-stubs = "^1.12.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
python_version = "3.10"
files = ["calendar_audit"]
plugins = ["mypy_django_plugin.main"]
exclude = ["^.*/migrations/.*\\.py$"]
disallow_untyped_defs = true
#disallow_any_unimported = true
no_implicit_optional = true
check_untyped_defs = true
warn_return_any = true
show_error_codes = true
warn_unused_ignores = true
ignore_missing_imports = true


[tool.django-stubs]
django_settings_module = "config.settings"

[tool.isort]
py_version = 310
profile = "black"
known_third_party = ["django", "environ", "google", "google_auth_oauthlib", "googleapiclient", "jwt", "requests"]
